# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /partner/orders:
    get:
      operationId: listPartnerOrders
      tags: [orders]
      summary: List orders belonging to the authenticated partner restaurant
      security:
        - bearerAuth: []
      parameters:
        - { name: page, in: query, schema: { type: number, minimum: 1, default: 1 } }
        - { name: pageSize, in: query, schema: { type: number, minimum: 1, maximum: 50, default: 20 } }
        - { name: status, in: query, schema: { type: string, enum: [pending, confirmed, preparing, ready, picked_up, delivered, cancelled] } }
      responses:
        "200":
          description: Restaurant orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PartnerOrderSummary"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Partner restaurant not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /partner/orders/{id}:
    get:
      operationId: getPartnerOrder
      tags: [orders]
      summary: Get an authenticated partner restaurant order
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Restaurant order detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PartnerOrderDetail"
        "400":
          description: Invalid order id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Order does not belong to this restaurant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /partner/orders/{id}/status:
    patch:
      operationId: updatePartnerOrderStatus
      tags: [orders]
      summary: Advance a restaurant-owned order through confirmation and preparation
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PartnerOrderStatusUpdate"
      responses:
        "200":
          description: Updated order state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DriverOrderStatusResult"
        "400":
          description: Invalid transition
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Order does not belong to this restaurant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Order not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
...
    OrderStatus:
      type: string
      enum: [pending, confirmed, preparing, ready, picked_up, delivered, cancelled]
    PaymentMethod:
      type: string
      enum: [cash, card]
    PaymentStatus:
      type: string
      enum: [pending, paid, failed, refunded]
...
    OrderAddon:
      type: object
      required: [name, price]
      properties:
        name:
          type: string
        price:
          type: number
    OrderLine:
      type: object
      required: [id, productId, name, quantity, unitPrice, lineTotal, addons]
      properties:
        id:
          type: number
        productId:
          type: number
        name:
          type: string
        variantName:
          type: ["string", "null"]
        quantity:
          type: number
        unitPrice:
          type: number
        lineTotal:
          type: number
        addons:
          type: array
          items:
            $ref: "#/components/schemas/OrderAddon"
    OrderTimelineEntry:
      type: object
      required: [status, at, label]
      properties:
        status:
          $ref: "#/components/schemas/OrderStatus"
        at:
          type: string
          format: date-time
        label:
          type: string
...
    PartnerOrderStatusUpdate:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [confirmed, preparing, ready]
    PartnerOrderSummary:
      type: object
      required: [id, code, customerName, customerPhone, branchName, status, paymentMethod, paymentStatus, total, createdAt]
      properties:
        id:
          type: number
        code:
          type: string
        customerName:
          type: ["string", "null"]
        customerPhone:
          type: ["string", "null"]
        branchName:
          type: ["string", "null"]
        status:
          $ref: "#/components/schemas/OrderStatus"
        paymentMethod:
          $ref: "#/components/schemas/PaymentMethod"
        paymentStatus:
          $ref: "#/components/schemas/PaymentStatus"
        total:
          type: number
        createdAt:
          type: string
          format: date-time
    PartnerOrderDetail:
      allOf:
        - $ref: "#/components/schemas/PartnerOrderSummary"
        - type: object
          required: [deliveryAddressText, subtotal, deliveryFee, notes, driverName, dispatchStatus, timeline, items]
          properties:
            deliveryAddressText:
              type: string
            subtotal:
              type: number
            deliveryFee:
              type: number
            notes:
              type: ["string", "null"]
            driverName:
              type: ["string", "null"]
            dispatchStatus:
              type: object
              required: [state, nextRetryAt, offerExpiresAt, reason]
              properties:
                state:
                  type: string
                  enum: [not_started, actively_offered, retry_scheduled, assigned, terminal]
                nextRetryAt:
                  type: ["string", "null"]
                  format: date-time
                offerExpiresAt:
                  type: ["string", "null"]
                  format: date-time
                reason:
                  type: ["string", "null"]
                  enum: [NO_FRESH_ELIGIBLE_DRIVER, BRANCH_LOCATION_UNAVAILABLE, OFFER_CONFLICT, null]
            timeline:
              type: array
              items:
                $ref: "#/components/schemas/OrderTimelineEntry"
            items:
              type: array
              items:
                $ref: "#/components/schemas/OrderLine"
    DriverOrderStatusResult:
      type: object
      required: [id, status, updatedAt]
      properties:
        id:
          type: number
        status:
          $ref: "#/components/schemas/OrderStatus"
        updatedAt:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
        code:
          type: string